add ProfilePage.java refactor model
