+++ src/test/java/com/demoqa/model/books/AddListOfBooksBodyModel.java
@@ -1,4 +1,4 @@
-package com.demoqa.model;
+package com.demoqa.model.books;
 
 import lombok.Data;
