+++ src/test/java/com/demoqa/tests/BookTests.java
@@ -1,16 +1,14 @@
 package com.demoqa.tests;
 
 import com.demoqa.api.BooksApi;
-import com.demoqa.model.DeleteBookResponseModel;
-import com.demoqa.model.LoginResponseModel;
+import com.demoqa.model.books.DeleteBookResponseModel;
+import com.demoqa.model.authorization.LoginResponseModel;
+import com.demoqa.pages.ProfilePage;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvFileSource;
-import org.openqa.selenium.Cookie;
 
-import static com.codeborne.selenide.Condition.disappear;
 import static com.codeborne.selenide.Selenide.$;
 import static com.codeborne.selenide.Selenide.open;
-import static com.codeborne.selenide.WebDriverRunner.getWebDriver;
 import static com.demoqa.api.LoginApi.successLogin;
 
 public class BookTests extends TestBase {
@@ -19,20 +17,19 @@ public class BookTests extends TestBase {
     @CsvFileSource(resources = "/testdata.csv")
     void deleteExistentBookTests(String userName, String password, String isbn) {
         BooksApi bookApi = new BooksApi();
+        ProfilePage  profilePage = new ProfilePage();
         LoginResponseModel login = successLogin(userName, password);
 
-        open("/favicon.ico");
-        getWebDriver().manage().addCookie(new Cookie("userID", login.getUserId()));
-        getWebDriver().manage().addCookie(new Cookie("token", login.getToken()));
-        getWebDriver().manage().addCookie(new Cookie("expires", login.getExpires()));
+        profilePage.addCookies(login);
 
         bookApi.deleteAllBooks(login.getUserId(), login.getToken());
         bookApi.addBook(login.getUserId(), login.getToken(), isbn);
 
         bookApi.deleteOneBook(login.getUserId(), login.getToken(), isbn);
 
-        open("/profile");
-        $("[id='see-book-Git Pocket Guide']").shouldBe(disappear);
+        profilePage
+                .openProfilePage()
+                .checkEmptyTable();
     }
 
     @ParameterizedTest(name = "Deleting non-existent book")
